requirements.txt
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
colorama==0.4.4
decorator==4.4.2
Django==3.1.4
django-rest-framework==0.1.0
djangorestframework==3.12.2
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
jsonschema==3.2.0
MarkupSafe==1.1.1
nbformat==5.0.8
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
Pygments==2.7.2
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
pyzmq==20.0.0
six==1.15.0
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
